+++ src/cscapi/storage.py
@@ -87,7 +87,13 @@ class SignalModel:
 
 class StorageInterface(ABC):
     @abstractmethod
-    def get_all_signals(self, limit: int, offset: int = 0) -> List[SignalModel]:
+    def get_all_signals(
+        self,
+        limit: int,
+        offset: int = 0,
+        sent: Optional[bool] = False,
+        is_failing: Optional[bool] = False,
+    ) -> List[SignalModel]:
         raise NotImplementedError
 
     @abstractmethod