+++ tests/test_client.py
@@ -216,7 +216,10 @@ class TestSendSignals:
         s1 = replace(mock_signals()[0], scenario="crowdsecurity/http-bf")
         s2 = mock_signals()[0]
         client.add_signals([s1, s2])
-        assert len(client.storage.get_all_signals(limit=1000)) == 2
+        assert (
+            len(client.storage.get_all_signals(limit=1000, sent=None, is_failing=None))
+            == 2
+        )
 
         assert client.storage.get_machine_by_id("test") is None