+++ CHANGELOG.md
@@ -16,7 +16,7 @@ functions provided by the `src/cscapi` folder.
 
 ### Changed
 
-- **Breaking change**: Modify `StorageInterface::get_all_signals` to accept a `limit` and `offset` arguments
+- **Breaking change**: Modify `StorageInterface::get_all_signals` to accept `limit`, `offset`, `sent` and `is_failing` arguments
 - **Breaking change**: Change `StorageInterface::delete_signals` signature to require a list of signal ids
 - **Breaking change**: Change `StorageInterface::delete_machines` signature to require a list of machine ids
 - **Breaking change**: `SQLStorage::get_all_signals` requires now a `limit` argument